refactor: make pricing.py the single source of truth for pricing
Move the PRICING table out of cli.py into a new pricing.py module and
inject it into the dashboard HTML at render time as JSON, so the
JavaScript copy can never drift from the Python one.

* pricing.py: PRICING dict + get_pricing/calc_cost helpers
* cli.py: import from pricing instead of defining its own copy
* dashboard.py: HTML_TEMPLATE carries /*__PRICING_JSON__*/ placeholder;
  new render_html() injects json.dumps(PRICING) at request time
* tests: TestPricingParity now parses the rendered HTML (post-injection)
  rather than the raw template

No behavior change — both the CLI cost calc and the dashboard JS see the
exact same numbers they did before. Adding a new model now means editing
exactly one place.

Closes #82
