+++ dashboard.py
@@ -7,6 +7,8 @@ import os
 import sqlite3
 from http.server import HTTPServer, BaseHTTPRequestHandler
 from pathlib import Path
+
+from pricing import PRICING
 from datetime import datetime
 
 DB_PATH = Path.home() / ".claude" / "usage.db"
@@ -418,17 +420,7 @@ function tzDisplayName(tzMode) {
 }
 
 // ── Pricing (Anthropic API, April 2026) ────────────────────────────────────
-const PRICING = {
-  'claude-opus-4-7':   { input:  5.00, output: 25.00, cache_write:  6.25, cache_read: 0.50 },
-  'claude-opus-4-6':   { input:  5.00, output: 25.00, cache_write:  6.25, cache_read: 0.50 },
-  'claude-opus-4-5':   { input:  5.00, output: 25.00, cache_write:  6.25, cache_read: 0.50 },
-  'claude-sonnet-4-7': { input:  3.00, output: 15.00, cache_write:  3.75, cache_read: 0.30 },
-  'claude-sonnet-4-6': { input:  3.00, output: 15.00, cache_write:  3.75, cache_read: 0.30 },
-  'claude-sonnet-4-5': { input:  3.00, output: 15.00, cache_write:  3.75, cache_read: 0.30 },
-  'claude-haiku-4-7':  { input:  1.00, output:  5.00, cache_write:  1.25, cache_read: 0.10 },
-  'claude-haiku-4-6':  { input:  1.00, output:  5.00, cache_write:  1.25, cache_read: 0.10 },
-  'claude-haiku-4-5':  { input:  1.00, output:  5.00, cache_write:  1.25, cache_read: 0.10 },
-};
+const PRICING = /*__PRICING_JSON__*/;
 
 function isBillable(model) {
   if (!model) return false;
@@ -1237,6 +1229,15 @@ scheduleAutoRefresh();
 """
 
 
+def render_html():
+    """Inject the Python PRICING table into the HTML so the JS table
+    can never drift from the Python one."""
+    return HTML_TEMPLATE.replace(
+        "/*__PRICING_JSON__*/",
+        json.dumps(PRICING),
+    ).encode("utf-8")
+
+
 class DashboardHandler(BaseHTTPRequestHandler):
     def log_message(self, format, *args):
         pass
@@ -1246,7 +1247,7 @@ class DashboardHandler(BaseHTTPRequestHandler):
             self.send_response(200)
             self.send_header("Content-Type", "text/html; charset=utf-8")
             self.end_headers()
-            self.wfile.write(HTML_TEMPLATE.encode("utf-8"))
+            self.wfile.write(render_html())
 
         elif self.path == "/api/data":
             data = get_dashboard_data()
